feat: ChessBoard 생성 전략 추가
