+++ src/main/java/chess/domain/generator/BoardGenerator.java
@@ -1,0 +1,10 @@
+package chess.domain.generator;
+
+import chess.domain.piece.Piece;
+import java.util.List;
+
+@FunctionalInterface
+public interface BoardGenerator {
+
+    List<List<Piece>> generate();
+}
